feat: BookDetailResponse 가격 타입 변경 - MapStruct int->BigDecimal

+++ src/main/java/com/nhnacademy/illuwa/d_book/book/mapper/BookResponseMapper.java
@@ -40,8 +40,8 @@ public interface BookResponseMapper {
                 book.getPublisher(),
                 book.getPublishedDate(),
                 book.getIsbn(),
-                book.getRegularPrice().intValue(),
-                book.getSalePrice().intValue(),
+                book.getRegularPrice(),
+                book.getSalePrice(),
                 book.getBookExtraInfo() != null && book.getBookExtraInfo().isGiftwrap(),
                 thumbnailUrl
         );
